+++ apps/app/src/components/PageEditor/markdown-table-util-for-editor.ts
@@ -2,12 +2,10 @@ import { EditorView } from '@codemirror/view';
 
 import MarkdownTable from '~/client/models/MarkdownTable';
 
-/**
- * Utility for markdown table
- */
-
 // https://regex101.com/r/7BN2fR/10
 const linePartOfTableRE = /^([^\r\n|]*)\|(([^\r\n|]*\|)+)$/;
+// https://regex101.com/r/1UuWBJ/3
+export const emptyLineOfTableRE = /^([^\r\n|]*)\|((\s*\|)+)$/;
 
 const curPos = (editor: EditorView): number => {
   return editor.state.selection.main.head;
